+++ __init__.py
@@ -199,106 +199,106 @@ def sna_update_active_mode_4A881(self, context):
     else:
         pass
 
-def sna_about_kiri_links_docs_3dgs_D02EC(layout_function, ):
-    box_12166 = layout_function.box()
-    box_12166.alert = False
-    box_12166.enabled = True
-    box_12166.active = True
-    box_12166.use_property_split = False
-    box_12166.use_property_decorate = False
-    box_12166.alignment = 'Expand'.upper()
-    box_12166.scale_x = 1.0
-    box_12166.scale_y = 1.0
-    if not True: box_12166.operator_context = "EXEC_DEFAULT"
-    col_CB7BF = box_12166.column(heading='', align=True)
-    col_CB7BF.alert = False
-    col_CB7BF.enabled = True
-    col_CB7BF.active = True
-    col_CB7BF.use_property_split = False
-    col_CB7BF.use_property_decorate = False
-    col_CB7BF.scale_x = 1.0
-    col_CB7BF.scale_y = 1.0
-    col_CB7BF.alignment = 'Expand'.upper()
-    col_CB7BF.operator_context = "INVOKE_DEFAULT" if True else "EXEC_DEFAULT"
-    box_D36BE = col_CB7BF.box()
-    box_D36BE.alert = False
-    box_D36BE.enabled = True
-    box_D36BE.active = True
-    box_D36BE.use_property_split = False
-    box_D36BE.use_property_decorate = False
-    box_D36BE.alignment = 'Center'.upper()
-    box_D36BE.scale_x = 1.0
-    box_D36BE.scale_y = 1.0
-    if not True: box_D36BE.operator_context = "EXEC_DEFAULT"
-    box_D36BE.label(text='About KIRI Engine', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'pointer-right-fill.svg')))
-    box_D36BE.template_icon(icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'Addon speel 2.png')), scale=10.0)
-    op = box_D36BE.operator('sna.dgs_render_launch_kiri_site_bf973', text='Learn More', icon_value=0, emboss=True, depress=False)
-    box_28D7F = col_CB7BF.box()
-    box_28D7F.alert = False
-    box_28D7F.enabled = True
-    box_28D7F.active = True
-    box_28D7F.use_property_split = False
-    box_28D7F.use_property_decorate = True
-    box_28D7F.alignment = 'Expand'.upper()
-    box_28D7F.scale_x = 1.0
-    box_28D7F.scale_y = 1.2000000476837158
-    if not True: box_28D7F.operator_context = "EXEC_DEFAULT"
-    split_F339F = box_28D7F.split(factor=0.30000001192092896, align=False)
-    split_F339F.alert = False
-    split_F339F.enabled = True
-    split_F339F.active = True
-    split_F339F.use_property_split = False
-    split_F339F.use_property_decorate = False
-    split_F339F.scale_x = 1.0
-    split_F339F.scale_y = 1.0
-    split_F339F.alignment = 'Expand'.upper()
-    if not True: split_F339F.operator_context = "EXEC_DEFAULT"
-    split_F339F.prop(bpy.context.scene.sna_dgs_scene_properties, 'show_tips', text='Tips', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'tips-one.svg')), emboss=True)
-    row_273F5 = split_F339F.row(heading='', align=False)
-    row_273F5.alert = False
-    row_273F5.enabled = True
-    row_273F5.active = True
-    row_273F5.use_property_split = False
-    row_273F5.use_property_decorate = False
-    row_273F5.scale_x = 1.0
-    row_273F5.scale_y = 1.0
-    row_273F5.alignment = 'Right'.upper()
-    row_273F5.operator_context = "INVOKE_DEFAULT" if True else "EXEC_DEFAULT"
-    op = row_273F5.operator('sna.dgs_render_open_documentation_3a04f', text='Documentation', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'documentation.svg')), emboss=True, depress=False)
-    op = row_273F5.operator('sna.dgs_render_open_tutorial_video_0684a', text='', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'video.svg')), emboss=True, depress=False)
-    box_DE3AB = col_CB7BF.box()
-    box_DE3AB.alert = False
-    box_DE3AB.enabled = True
-    box_DE3AB.active = True
-    box_DE3AB.use_property_split = False
-    box_DE3AB.use_property_decorate = False
-    box_DE3AB.alignment = 'Expand'.upper()
-    box_DE3AB.scale_x = 1.0
-    box_DE3AB.scale_y = 1.0
-    if not True: box_DE3AB.operator_context = "EXEC_DEFAULT"
-    row_7D004 = box_DE3AB.row(heading='', align=False)
-    row_7D004.alert = False
-    row_7D004.enabled = True
-    row_7D004.active = True
-    row_7D004.use_property_split = False
-    row_7D004.use_property_decorate = False
-    row_7D004.scale_x = 1.0
-    row_7D004.scale_y = 1.2000000476837158
-    row_7D004.alignment = 'Expand'.upper()
-    row_7D004.operator_context = "INVOKE_DEFAULT" if True else "EXEC_DEFAULT"
-    row_7D004.label(text='Get More Addons', icon_value=0)
-    split_649B4 = row_7D004.split(factor=0.5, align=False)
-    split_649B4.alert = False
-    split_649B4.enabled = True
-    split_649B4.active = True
-    split_649B4.use_property_split = False
-    split_649B4.use_property_decorate = False
-    split_649B4.scale_x = 1.0
-    split_649B4.scale_y = 1.0
-    split_649B4.alignment = 'Expand'.upper()
-    if not True: split_649B4.operator_context = "EXEC_DEFAULT"
-    op = split_649B4.operator('sna.dgs_render_launch_superhive_store_08f23', text='', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'SuperHive Logo White.png')), emboss=True, depress=False)
-    op = split_649B4.operator('sna.dgs_render_launch_kiri_blender_addons_page_9d58c', text='', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'kiriengine blender addon icon color.svg')), emboss=True, depress=False)
+# def sna_about_kiri_links_docs_3dgs_D02EC(layout_function, ):
+#     box_12166 = layout_function.box()
+#     box_12166.alert = False
+#     box_12166.enabled = True
+#     box_12166.active = True
+#     box_12166.use_property_split = False
+#     box_12166.use_property_decorate = False
+#     box_12166.alignment = 'Expand'.upper()
+#     box_12166.scale_x = 1.0
+#     box_12166.scale_y = 1.0
+#     if not True: box_12166.operator_context = "EXEC_DEFAULT"
+#     col_CB7BF = box_12166.column(heading='', align=True)
+#     col_CB7BF.alert = False
+#     col_CB7BF.enabled = True
+#     col_CB7BF.active = True
+#     col_CB7BF.use_property_split = False
+#     col_CB7BF.use_property_decorate = False
+#     col_CB7BF.scale_x = 1.0
+#     col_CB7BF.scale_y = 1.0
+#     col_CB7BF.alignment = 'Expand'.upper()
+#     col_CB7BF.operator_context = "INVOKE_DEFAULT" if True else "EXEC_DEFAULT"
+#     box_D36BE = col_CB7BF.box()
+#     box_D36BE.alert = False
+#     box_D36BE.enabled = True
+#     box_D36BE.active = True
+#     box_D36BE.use_property_split = False
+#     box_D36BE.use_property_decorate = False
+#     box_D36BE.alignment = 'Center'.upper()
+#     box_D36BE.scale_x = 1.0
+#     box_D36BE.scale_y = 1.0
+#     if not True: box_D36BE.operator_context = "EXEC_DEFAULT"
+#     box_D36BE.label(text='About KIRI Engine', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'pointer-right-fill.svg')))
+#     box_D36BE.template_icon(icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'Addon speel 2.png')), scale=10.0)
+#     op = box_D36BE.operator('sna.dgs_render_launch_kiri_site_bf973', text='Learn More', icon_value=0, emboss=True, depress=False)
+#     box_28D7F = col_CB7BF.box()
+#     box_28D7F.alert = False
+#     box_28D7F.enabled = True
+#     box_28D7F.active = True
+#     box_28D7F.use_property_split = False
+#     box_28D7F.use_property_decorate = True
+#     box_28D7F.alignment = 'Expand'.upper()
+#     box_28D7F.scale_x = 1.0
+#     box_28D7F.scale_y = 1.2000000476837158
+#     if not True: box_28D7F.operator_context = "EXEC_DEFAULT"
+#     split_F339F = box_28D7F.split(factor=0.30000001192092896, align=False)
+#     split_F339F.alert = False
+#     split_F339F.enabled = True
+#     split_F339F.active = True
+#     split_F339F.use_property_split = False
+#     split_F339F.use_property_decorate = False
+#     split_F339F.scale_x = 1.0
+#     split_F339F.scale_y = 1.0
+#     split_F339F.alignment = 'Expand'.upper()
+#     if not True: split_F339F.operator_context = "EXEC_DEFAULT"
+#     split_F339F.prop(bpy.context.scene.sna_dgs_scene_properties, 'show_tips', text='Tips', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'tips-one.svg')), emboss=True)
+#     row_273F5 = split_F339F.row(heading='', align=False)
+#     row_273F5.alert = False
+#     row_273F5.enabled = True
+#     row_273F5.active = True
+#     row_273F5.use_property_split = False
+#     row_273F5.use_property_decorate = False
+#     row_273F5.scale_x = 1.0
+#     row_273F5.scale_y = 1.0
+#     row_273F5.alignment = 'Right'.upper()
+#     row_273F5.operator_context = "INVOKE_DEFAULT" if True else "EXEC_DEFAULT"
+#     op = row_273F5.operator('sna.dgs_render_open_documentation_3a04f', text='Documentation', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'documentation.svg')), emboss=True, depress=False)
+#     op = row_273F5.operator('sna.dgs_render_open_tutorial_video_0684a', text='', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'video.svg')), emboss=True, depress=False)
+#     box_DE3AB = col_CB7BF.box()
+#     box_DE3AB.alert = False
+#     box_DE3AB.enabled = True
+#     box_DE3AB.active = True
+#     box_DE3AB.use_property_split = False
+#     box_DE3AB.use_property_decorate = False
+#     box_DE3AB.alignment = 'Expand'.upper()
+#     box_DE3AB.scale_x = 1.0
+#     box_DE3AB.scale_y = 1.0
+#     if not True: box_DE3AB.operator_context = "EXEC_DEFAULT"
+#     row_7D004 = box_DE3AB.row(heading='', align=False)
+#     row_7D004.alert = False
+#     row_7D004.enabled = True
+#     row_7D004.active = True
+#     row_7D004.use_property_split = False
+#     row_7D004.use_property_decorate = False
+#     row_7D004.scale_x = 1.0
+#     row_7D004.scale_y = 1.2000000476837158
+#     row_7D004.alignment = 'Expand'.upper()
+#     row_7D004.operator_context = "INVOKE_DEFAULT" if True else "EXEC_DEFAULT"
+#     row_7D004.label(text='Get More Addons', icon_value=0)
+#     split_649B4 = row_7D004.split(factor=0.5, align=False)
+#     split_649B4.alert = False
+#     split_649B4.enabled = True
+#     split_649B4.active = True
+#     split_649B4.use_property_split = False
+#     split_649B4.use_property_decorate = False
+#     split_649B4.scale_x = 1.0
+#     split_649B4.scale_y = 1.0
+#     split_649B4.alignment = 'Expand'.upper()
+#     if not True: split_649B4.operator_context = "EXEC_DEFAULT"
+#     op = split_649B4.operator('sna.dgs_render_launch_superhive_store_08f23', text='', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'SuperHive Logo White.png')), emboss=True, depress=False)
+#     op = split_649B4.operator('sna.dgs_render_launch_kiri_blender_addons_page_9d58c', text='', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'kiriengine blender addon icon color.svg')), emboss=True, depress=False)
 
 
 
@@ -450,24 +450,24 @@ def sna_about_kiri_links_docs_3dgs_D02EC(layout_function, ):
 #             box_DC5A2.prop(bpy.context.view_layer.objects.active.sna_dgs_object_properties, 'cam_update', text='Use Active Camera', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'camera.svg')), emboss=True, toggle=True)
 #
 
-class SNA_OT_Dgs_Render_Align_Active_To_X_Axis_6Ae0E(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_align_active_to_x_axis_6ae0e"
-    bl_label = "3DGS Render: Align Active To X Axis"
-    bl_description = "Updates the 3DGS_Render modifier once to the X axis for the active object."
-    bl_options = {"REGISTER", "UNDO"}
-
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
-
-    def execute(self, context):
-        sna_align_active_values_to_x_4CE1F()
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        return self.execute(context)
+# class SNA_OT_Dgs_Render_Align_Active_To_X_Axis_6Ae0E(bpy.types.Operator):
+#     bl_idname = "sna.dgs_render_align_active_to_x_axis_6ae0e"
+#     bl_label = "3DGS Render: Align Active To X Axis"
+#     bl_description = "Updates the 3DGS_Render modifier once to the X axis for the active object."
+#     bl_options = {"REGISTER", "UNDO"}
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         sna_align_active_values_to_x_4CE1F()
+#         return {"FINISHED"}
+#
+#     def invoke(self, context, event):
+#         return self.execute(context)
 
 # class SNA_OT_Dgs_Render_Align_Active_To_Z_Axis_1E184(bpy.types.Operator):
 #     bl_idname = "sna.dgs_render_align_active_to_z_axis_1e184"
@@ -782,29 +782,29 @@ class SNA_OT_Dgs_Render_Align_Active_To_X_Axis_6Ae0E(bpy.types.Operator):
 #         row_07DA9.label(text='Animate Modifier Is Missing', icon_value=0)
 #         op = row_07DA9.operator('sna.dgs_render_add_animate_modifier_39c55', text='', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'plus-circle.svg')), emboss=True, depress=False)
 
-class SNA_OT_Dgs_Render_Add_Animate_Modifier_39C55(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_add_animate_modifier_39c55"
-    bl_label = "3DGS Render: Add Animate Modifier"
-    bl_description = "Adds a 3DGS animate modifier to the active object."
-    bl_options = {"REGISTER", "UNDO"}
-
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
-
-    def execute(self, context):
-        if (bpy.context.view_layer.objects.active.type == 'MESH' or bpy.context.view_layer.objects.active.type == 'CURVE'):
-            created_modifier_0_3280f = sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Animate_GN', 'KIRI_3DGS_Animate_GN', bpy.context.view_layer.objects.active)
-            if (len(bpy.context.view_layer.objects.active.modifiers) > 0):
-                pass
-        else:
-            self.report({'INFO'}, message='The Active Object is not a mesh or curve object.')
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        return self.execute(context)
+# class SNA_OT_Dgs_Render_Add_Animate_Modifier_39C55(bpy.types.Operator):
+#     bl_idname = "sna.dgs_render_add_animate_modifier_39c55"
+#     bl_label = "3DGS Render: Add Animate Modifier"
+#     bl_description = "Adds a 3DGS animate modifier to the active object."
+#     bl_options = {"REGISTER", "UNDO"}
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         if (bpy.context.view_layer.objects.active.type == 'MESH' or bpy.context.view_layer.objects.active.type == 'CURVE'):
+#             created_modifier_0_3280f = sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Animate_GN', 'KIRI_3DGS_Animate_GN', bpy.context.view_layer.objects.active)
+#             if (len(bpy.context.view_layer.objects.active.modifiers) > 0):
+#                 pass
+#         else:
+#             self.report({'INFO'}, message='The Active Object is not a mesh or curve object.')
+#         return {"FINISHED"}
+#
+#     def invoke(self, context, event):
+#         return self.execute(context)
 
 
 def sna_add_to_view3d_mt_object_apply_F9005(self, context):
@@ -2636,48 +2636,48 @@ class SNA_OT_Dgs_Render_Generate_Hq_Object_55455(bpy.types.Operator):
         return context.window_manager.invoke_props_dialog(self, width=500)
 
 
-class SNA_OT_Dgs_Render_Disable_Hq_Overlap_34678(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_disable_hq_overlap_34678"
-    bl_label = "3DGS Render: Disable HQ Overlap"
-    bl_description = ""
-    bl_options = {"REGISTER", "UNDO"}
-
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
-
-    def execute(self, context):
-        bpy.data.objects.remove(object=bpy.data.objects['KIRI_HQ_Merged_Object'], do_unlink=True, do_id_user=True, do_ui_user=True, )
-        if property_exists("bpy.data.collections['3DGS_HQ_Object']", globals(), locals()):
-            bpy.data.collections.remove(collection=bpy.data.collections['3DGS_HQ_Object'], do_unlink=True, do_id_user=True, do_ui_user=True, )
-        if property_exists("bpy.data.collections['3DGS_LQ_Objects']", globals(), locals()):
-            for i_EF373 in range(len(bpy.data.collections['3DGS_LQ_Objects'].all_objects)):
-                bpy.data.collections['3DGS_LQ_Objects'].all_objects[i_EF373].hide_viewport = False
-                bpy.data.collections['3DGS_LQ_Objects'].all_objects[i_EF373].hide_render = False
-        if bpy.context and bpy.context.screen:
-            for a in bpy.context.screen.areas:
-                a.tag_redraw()
-        return {"FINISHED"}
-
-    def draw(self, context):
-        layout = self.layout
-        box_2D485 = layout.box()
-        box_2D485.alert = False
-        box_2D485.enabled = True
-        box_2D485.active = True
-        box_2D485.use_property_split = False
-        box_2D485.use_property_decorate = False
-        box_2D485.alignment = 'Expand'.upper()
-        box_2D485.scale_x = 1.0
-        box_2D485.scale_y = 1.0
-        if not True: box_2D485.operator_context = "EXEC_DEFAULT"
-        box_2D485.label(text='HQ Object Found In Scene', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'tips-one.svg')))
-        box_2D485.label(text='        Remove It?', icon_value=0)
-
-    def invoke(self, context, event):
-        return context.window_manager.invoke_props_dialog(self, width=300)
+# class SNA_OT_Dgs_Render_Disable_Hq_Overlap_34678(bpy.types.Operator):
+#     bl_idname = "sna.dgs_render_disable_hq_overlap_34678"
+#     bl_label = "3DGS Render: Disable HQ Overlap"
+#     bl_description = ""
+#     bl_options = {"REGISTER", "UNDO"}
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         bpy.data.objects.remove(object=bpy.data.objects['KIRI_HQ_Merged_Object'], do_unlink=True, do_id_user=True, do_ui_user=True, )
+#         if property_exists("bpy.data.collections['3DGS_HQ_Object']", globals(), locals()):
+#             bpy.data.collections.remove(collection=bpy.data.collections['3DGS_HQ_Object'], do_unlink=True, do_id_user=True, do_ui_user=True, )
+#         if property_exists("bpy.data.collections['3DGS_LQ_Objects']", globals(), locals()):
+#             for i_EF373 in range(len(bpy.data.collections['3DGS_LQ_Objects'].all_objects)):
+#                 bpy.data.collections['3DGS_LQ_Objects'].all_objects[i_EF373].hide_viewport = False
+#                 bpy.data.collections['3DGS_LQ_Objects'].all_objects[i_EF373].hide_render = False
+#         if bpy.context and bpy.context.screen:
+#             for a in bpy.context.screen.areas:
+#                 a.tag_redraw()
+#         return {"FINISHED"}
+#
+#     def draw(self, context):
+#         layout = self.layout
+#         box_2D485 = layout.box()
+#         box_2D485.alert = False
+#         box_2D485.enabled = True
+#         box_2D485.active = True
+#         box_2D485.use_property_split = False
+#         box_2D485.use_property_decorate = False
+#         box_2D485.alignment = 'Expand'.upper()
+#         box_2D485.scale_x = 1.0
+#         box_2D485.scale_y = 1.0
+#         if not True: box_2D485.operator_context = "EXEC_DEFAULT"
+#         box_2D485.label(text='HQ Object Found In Scene', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'tips-one.svg')))
+#         box_2D485.label(text='        Remove It?', icon_value=0)
+#
+#     def invoke(self, context, event):
+#         return context.window_manager.invoke_props_dialog(self, width=300)
 
 
 # def sna_dgs_render__main_function_menu_019C7(layout_function, ):
@@ -2732,24 +2732,24 @@ class SNA_OT_Dgs_Render_Disable_Hq_Overlap_34678(bpy.types.Operator):
 #         pass
 
 
-class SNA_OT_Dgs_Render_Update_Enabled_3Dgs_Objects_6D7F4(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_update_enabled_3dgs_objects_6d7f4"
-    bl_label = "3DGS Render: Update Enabled 3DGS Objects"
-    bl_description = "Updates all enabled 3DGS objects faces to the current view."
-    bl_options = {"REGISTER", "UNDO"}
-
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
-
-    def execute(self, context):
-        sna_update_camera_single_time_9EF18()
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        return self.execute(context)
+# class SNA_OT_Dgs_Render_Update_Enabled_3Dgs_Objects_6D7F4(bpy.types.Operator):
+#     bl_idname = "sna.dgs_render_update_enabled_3dgs_objects_6d7f4"
+#     bl_label = "3DGS Render: Update Enabled 3DGS Objects"
+#     bl_description = "Updates all enabled 3DGS objects faces to the current view."
+#     bl_options = {"REGISTER", "UNDO"}
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         sna_update_camera_single_time_9EF18()
+#         return {"FINISHED"}
+#
+#     def invoke(self, context, event):
+#         return self.execute(context)
 
 
 # def sna_update_camera_single_time_9EF18():
@@ -4518,82 +4518,82 @@ class SNA_OT_Dgs_Render_Update_Enabled_3Dgs_Objects_6D7F4(bpy.types.Operator):
 #         sna_attribute_adjust_properties_2C323(layout_function, )
 #
 
-class SNA_OT_Dgs_Render_Refresh_Scene_C0B35(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_refresh_scene_c0b35"
-    bl_label = "3DGS Render: Refresh Scene"
-    bl_description = ""
-    bl_options = {"REGISTER", "UNDO"}
+# class SNA_OT_Dgs_Render_Refresh_Scene_C0B35(bpy.types.Operator):
+#     bl_idname = "sna.dgs_render_refresh_scene_c0b35"
+#     bl_label = "3DGS Render: Refresh Scene"
+#     bl_description = ""
+#     bl_options = {"REGISTER", "UNDO"}
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         if (bpy.context.scene.sna_dgs_scene_properties.r2_update_type == 'Single Time'):
+#             pass
+#         else:
+#             bpy.context.scene.frame_current = bpy.context.scene.frame_start
+#         if bpy.context and bpy.context.screen:
+#             for a in bpy.context.screen.areas:
+#                 a.tag_redraw()
+#         bpy.context.scene.sna_dgs_scene_properties.r2_interval_stop = (bpy.context.scene.sna_dgs_scene_properties.r2_update_type == 'Single Time')
+#
+#         def delayed_09B50():
+#             for i_75739 in range(len(bpy.context.scene.objects)):
+#                 if (property_exists("bpy.context.scene.objects[i_75739].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_75739].modifiers):
+#                     bpy.context.scene.objects[i_75739].hide_viewport = False
+#             sna_c2_refresh_all_4D367((not bpy.context.scene.sna_dgs_scene_properties.r2_selected), bpy.context.scene.sna_dgs_scene_properties.r2_transforms, True)
+#             sna_shader_system_A4AED()
+#             sna_texture_creation_FD1B2()
+#             sna_viewport_render_A3941()
+#             for i_07D59 in range(len(bpy.context.scene.objects)):
+#                 if ((property_exists("bpy.context.scene.objects[i_07D59].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_07D59].modifiers) and '3DGS_Mesh_Type' in bpy.context.scene.objects[i_07D59]):
+#                     if (bpy.context.scene.objects[i_07D59]['3DGS_Mesh_Type'] == 'face'):
+#                         bpy.context.scene.objects[i_07D59].hide_viewport = True
+#             if bpy.context.scene.sna_dgs_scene_properties.r2_interval_stop:
+#                 return None
+#             return bpy.context.scene.sna_dgs_scene_properties.r2_interval
+#         bpy.app.timers.register(delayed_09B50, first_interval=0.0)
+#         for i_BD2F9 in range(len(bpy.context.scene.objects)):
+#             if (property_exists("bpy.context.scene.objects[i_BD2F9].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_BD2F9].modifiers):
+#                 bpy.context.scene.objects[i_BD2F9].hide_viewport = False
+#         sna_c2_refresh_all_4D367((not bpy.context.scene.sna_dgs_scene_properties.r2_selected), bpy.context.scene.sna_dgs_scene_properties.r2_transforms, True)
+#         sna_shader_system_A4AED()
+#         sna_texture_creation_FD1B2()
+#         sna_viewport_render_A3941()
+#         for i_D854A in range(len(bpy.context.scene.objects)):
+#             if ((property_exists("bpy.context.scene.objects[i_D854A].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_D854A].modifiers) and '3DGS_Mesh_Type' in bpy.context.scene.objects[i_D854A]):
+#                 if (bpy.context.scene.objects[i_D854A]['3DGS_Mesh_Type'] == 'face'):
+#                     bpy.context.scene.objects[i_D854A].hide_viewport = True
+#         return {"FINISHED"}
+#
+#     def invoke(self, context, event):
+#         return self.execute(context)
 
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
 
-    def execute(self, context):
-        if (bpy.context.scene.sna_dgs_scene_properties.r2_update_type == 'Single Time'):
-            pass
-        else:
-            bpy.context.scene.frame_current = bpy.context.scene.frame_start
-        if bpy.context and bpy.context.screen:
-            for a in bpy.context.screen.areas:
-                a.tag_redraw()
-        bpy.context.scene.sna_dgs_scene_properties.r2_interval_stop = (bpy.context.scene.sna_dgs_scene_properties.r2_update_type == 'Single Time')
-
-        def delayed_09B50():
-            for i_75739 in range(len(bpy.context.scene.objects)):
-                if (property_exists("bpy.context.scene.objects[i_75739].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_75739].modifiers):
-                    bpy.context.scene.objects[i_75739].hide_viewport = False
-            sna_c2_refresh_all_4D367((not bpy.context.scene.sna_dgs_scene_properties.r2_selected), bpy.context.scene.sna_dgs_scene_properties.r2_transforms, True)
-            sna_shader_system_A4AED()
-            sna_texture_creation_FD1B2()
-            sna_viewport_render_A3941()
-            for i_07D59 in range(len(bpy.context.scene.objects)):
-                if ((property_exists("bpy.context.scene.objects[i_07D59].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_07D59].modifiers) and '3DGS_Mesh_Type' in bpy.context.scene.objects[i_07D59]):
-                    if (bpy.context.scene.objects[i_07D59]['3DGS_Mesh_Type'] == 'face'):
-                        bpy.context.scene.objects[i_07D59].hide_viewport = True
-            if bpy.context.scene.sna_dgs_scene_properties.r2_interval_stop:
-                return None
-            return bpy.context.scene.sna_dgs_scene_properties.r2_interval
-        bpy.app.timers.register(delayed_09B50, first_interval=0.0)
-        for i_BD2F9 in range(len(bpy.context.scene.objects)):
-            if (property_exists("bpy.context.scene.objects[i_BD2F9].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_BD2F9].modifiers):
-                bpy.context.scene.objects[i_BD2F9].hide_viewport = False
-        sna_c2_refresh_all_4D367((not bpy.context.scene.sna_dgs_scene_properties.r2_selected), bpy.context.scene.sna_dgs_scene_properties.r2_transforms, True)
-        sna_shader_system_A4AED()
-        sna_texture_creation_FD1B2()
-        sna_viewport_render_A3941()
-        for i_D854A in range(len(bpy.context.scene.objects)):
-            if ((property_exists("bpy.context.scene.objects[i_D854A].modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.scene.objects[i_D854A].modifiers) and '3DGS_Mesh_Type' in bpy.context.scene.objects[i_D854A]):
-                if (bpy.context.scene.objects[i_D854A]['3DGS_Mesh_Type'] == 'face'):
-                    bpy.context.scene.objects[i_D854A].hide_viewport = True
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        return self.execute(context)
-
-
-class SNA_OT_Dgs_Render_Create_Proxy_From_Mesh_D5B41(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_create_proxy_from_mesh_d5b41"
-    bl_label = "3DGS Render: Create Proxy From Mesh"
-    bl_description = ""
-    bl_options = {"REGISTER", "UNDO"}
-
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
-
-    def execute(self, context):
-        sna_b2_load_from_blender_object_F0CCB(bpy.context.view_layer.objects.active.name + 'Splat_Proxy')
-        sna_shader_system_A4AED()
-        sna_texture_creation_FD1B2()
-        sna_viewport_render_A3941()
-        return {"FINISHED"}
-
-    def invoke(self, context, event):
-        return self.execute(context)
+# class SNA_OT_Dgs_Render_Create_Proxy_From_Mesh_D5B41(bpy.types.Operator):
+#     bl_idname = "sna.dgs_render_create_proxy_from_mesh_d5b41"
+#     bl_label = "3DGS Render: Create Proxy From Mesh"
+#     bl_description = ""
+#     bl_options = {"REGISTER", "UNDO"}
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         sna_b2_load_from_blender_object_F0CCB(bpy.context.view_layer.objects.active.name + 'Splat_Proxy')
+#         sna_shader_system_A4AED()
+#         sna_texture_creation_FD1B2()
+#         sna_viewport_render_A3941()
+#         return {"FINISHED"}
+#
+#     def invoke(self, context, event):
+#         return self.execute(context)
 
 
 class SNA_OT_Dgs_Render_Clean_Up_Advanced_Render_Scene_09450(bpy.types.Operator):
@@ -4826,233 +4826,233 @@ def sna_align_active_values_to_z_7B9ED():
             a.tag_redraw()
 
 
-class SNA_OT_Dgs_Render_Import_Ply_E0A3A(bpy.types.Operator, ImportHelper):
-    bl_idname = "sna.dgs_render_import_ply_e0a3a"
-    bl_label = "3DGS Render: Import PLY"
-    bl_description = "Imports a .PLY file and adds 3DGS modifiers and attributes."
-    bl_options = {"REGISTER", "UNDO"}
-    filter_glob: bpy.props.StringProperty( default='*.ply', options={'HIDDEN'} )
-
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
-
-    def execute(self, context):
-        ply_import_path = self.filepath
-        output_object = None
-        attributes_missing = None
-        message = None
-        import os
-        # Input file path
-        #ply_import_path = "C:\\Users\\joe and pig\\Documents\\Flamingo.ply"  # Example path; update as needed
-        # Check if the file path is provided and exists
-        if not ply_import_path or not os.path.exists(ply_import_path):
-            print("Error: No file path provided or file not found.")
-            attributes_missing = True
-            message = f"Error: No file path provided or file not found at {ply_import_path}."
-            output_object = None
-        else:
-            try:
-                # Import the PLY file using Blender's native importer (Blender 4.0+)
-                try:
-                    bpy.ops.wm.ply_import(filepath=ply_import_path)
-                except AttributeError:
-                    print("Error: PLY importer operator 'bpy.ops.wm.ply_import' not found. Ensure Blender 4.0 or later is used.")
-                    attributes_missing = True
-                    message = f"Error: PLY importer operator not found for {os.path.basename(ply_import_path)}. Ensure Blender 4.0 or later is used."
-                    output_object = None
-                else:
-                    # Get the last imported object (assuming it's the PLY mesh)
-                    imported_objects = [obj for obj in bpy.context.scene.objects if obj.select_get()]
-                    if not imported_objects:
-                        print("Error: No objects were imported from the PLY file.")
-                        attributes_missing = True
-                        message = f"Error: No objects were imported from the PLY file {os.path.basename(ply_import_path)}."
-                        output_object = None
-                    else:
-                        # Use the first imported object (assuming it's a mesh)
-                        obj = imported_objects[0]
-                        if obj.type != 'MESH':
-                            print("Error: Imported object is not a mesh.")
-                            attributes_missing = True
-                            message = f"Error: Imported object from {os.path.basename(ply_import_path)} is not a mesh."
-                            output_object = None
-                        else:
-                            # Get the mesh data
-                            mesh = obj.data
-                            # List of required 3DGS attributes
-                            required_attributes = ['f_dc_0', 'f_dc_1', 'f_dc_2', 'opacity', 'scale_0', 'scale_1', 'scale_2', 'rot_0', 'rot_1', 'rot_2', 'rot_3', 'f_rest_0']
-                            # Check if all required attributes exist on the mesh
-                            attributes_missing = False
-                            missing_attrs = []
-                            for attr_name in required_attributes:
-                                if attr_name not in mesh.attributes:
-                                    attributes_missing = True
-                                    missing_attrs.append(attr_name)
-                            # Set the message and output object based on success or failure
-                            if attributes_missing:
-                                message = f"Error: One or more required 3DGS attributes ({', '.join(missing_attrs)}) are missing on {obj.name}. Check that .ply is a 3DGS scan."
-                                print(message)
-                                output_object = None
-                            else:
-                                message = f"Successfully imported PLY file {os.path.basename(ply_import_path)} and all required 3DGS attributes ({', '.join(required_attributes)}) are present on {obj.name}."
-                                print(message)
-                                output_object = obj
-            except Exception as e:
-                print(f"Error importing PLY file: {e}")
-                attributes_missing = True
-                message = f"Error importing PLY file {os.path.basename(ply_import_path)}: {str(e)}"
-                output_object = None
-        # The variables are now available for use
-        # You can access 'attributes_missing' (boolean), 'message' (string), and 'output_object' (bpy.types.Object or None) in other scripts or logic
-        if (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces'):
-            bpy.context.view_layer.objects.active['3DGS_Mesh_Type'] = 'face'
-            sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Store_Origpos_GN', 'KIRI_3DGS_Store_Origpos_GN', bpy.context.view_layer.objects.active)
-            bpy.ops.object.modifier_apply('INVOKE_DEFAULT', modifier='KIRI_3DGS_Store_Origpos_GN')
-            sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Render_GN', 'KIRI_3DGS_Render_GN', bpy.context.view_layer.objects.active)
-            sna_align_active_values_to_x_4CE1F()
-            bpy.ops.object.modifier_apply('INVOKE_DEFAULT', modifier='KIRI_3DGS_Render_GN')
-        else:
-            bpy.context.view_layer.objects.active['3DGS_Mesh_Type'] = 'vert'
-        import numpy as np
-        # Constants from your script
-        SH_0 = 0.28209479177387814
-        # Get the active object
-        obj = bpy.context.active_object
-        if not obj or obj.type != 'MESH':
-            print("Error: No active mesh object selected.")
-        else:
-            # Get the mesh data
-            mesh = obj.data
-            # Check if the required attributes exist
-            if not all(attr.name in mesh.attributes for attr in mesh.attributes if attr.name in ['f_dc_0', 'f_dc_1', 'f_dc_2', 'opacity']):
-                print("Error: Required attributes (f_dc_0, f_dc_1, f_dc_2, opacity) not found on the mesh.")
-            else:
-                # Get the number of points (vertices)
-                point_count = len(mesh.vertices)
-                expected_length = point_count * 4  # Each point has 1 RGBA set (4 values) for both Col and KIRI_3DGS_Paint
-                # Extract data from attributes (assuming they are on POINT domain)
-                f_dc_0_data = np.array([v.value for v in mesh.attributes['f_dc_0'].data])
-                f_dc_1_data = np.array([v.value for v in mesh.attributes['f_dc_1'].data])
-                f_dc_2_data = np.array([v.value for v in mesh.attributes['f_dc_2'].data])
-                opacity_data = np.array([v.value for v in mesh.attributes['opacity'].data])
-                # Debug: Check lengths of input attribute data
-                print(f"Number of points (vertices): {point_count}")
-                print(f"Length of f_dc_0_data: {len(f_dc_0_data)}")
-                print(f"Length of f_dc_1_data: {len(f_dc_1_data)}")
-                print(f"Length of f_dc_2_data: {len(f_dc_2_data)}")
-                print(f"Length of opacity_data: {len(opacity_data)}")
-                # Verify that attribute data lengths match the number of points
-                if not (len(f_dc_0_data) == len(f_dc_1_data) == len(f_dc_2_data) == len(opacity_data) == point_count):
-                    print("Error: Mismatch in attribute data lengths. Expected length matches point_count.")
-                else:
-                    # Calculate RGB and Alpha for each point (same calculation for both Col and KIRI_3DGS_Paint)
-                    color_data = []  # For both Col and KIRI_3DGS_Paint, one RGBA per point
-                    for i in range(point_count):
-                        # Calculate RGB (matching your script)
-                        R = (f_dc_0_data[i] * SH_0 + 0.5)
-                        G = (f_dc_1_data[i] * SH_0 + 0.5)
-                        B = (f_dc_2_data[i] * SH_0 + 0.5)
-                        # Calculate Alpha (using sigmoid if opacity is in log-space, or raw if [0, 1])
-                        # Here, we assume opacity is in log-space (logits) as in your script
-                        log_opacity = opacity_data[i]
-                        A = 1 / (1 + np.exp(-log_opacity))
-                        # Ensure values are in [0, 1]
-                        R = max(0.0, min(1.0, R))
-                        G = max(0.0, min(1.0, G))
-                        B = max(0.0, min(1.0, B))
-                        A = max(0.0, min(1.0, A))
-                        # Add RGBA for both Col and KIRI_3DGS_Paint (one set per point)
-                        color_data.extend([R, G, B, A])
-                    # Debug: Check calculated data length
-                    print(f"Length of color_data (Col and KIRI_3DGS_Paint): {len(color_data)}")
-                    print(f"Expected length for POINT domain: {expected_length}")
-                    # Verify data length matches expectation
-                    if len(color_data) != expected_length:
-                        print(f"Error: Array length mismatch (expected {expected_length}, got {len(color_data)})")
-                    else:
-                        # Create or update the Col attribute on the point domain
-                        if 'Col' in mesh.attributes:
-                            mesh.attributes.remove(mesh.attributes['Col'])
-                        col_attr = mesh.attributes.new(name="Col", type='FLOAT_COLOR', domain='POINT')
-                        col_attr.data.foreach_set("color", color_data)
-                        # Create or update the KIRI_3DGS_Paint attribute on the point domain
-                        if 'KIRI_3DGS_Paint' in mesh.attributes:
-                            mesh.attributes.remove(mesh.attributes['KIRI_3DGS_Paint'])
-                        paint_attr = mesh.attributes.new(name="KIRI_3DGS_Paint", type='FLOAT_COLOR', domain='POINT')
-                        paint_attr.data.foreach_set("color", color_data)
-                        # Set KIRI_3DGS_Paint as the active color attribute
-                        mesh.color_attributes.active_color = paint_attr
-                        print(f"Created Col attribute on {obj.name} with {point_count} points.")
-                        print(f"Created KIRI_3DGS_Paint attribute on {obj.name} with {point_count} color values on the POINT domain.")
-        sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Render_GN', 'KIRI_3DGS_Render_GN', bpy.context.view_layer.objects.active)
-        sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Sorter_GN', 'KIRI_3DGS_Sorter_GN', bpy.context.view_layer.objects.active)
-        sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Adjust_Colour_And_Material', 'KIRI_3DGS_Adjust_Colour_And_Material', bpy.context.view_layer.objects.active)
-        sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Write F_DC_And_Merge', 'KIRI_3DGS_Write F_DC_And_Merge', bpy.context.view_layer.objects.active)
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN'].show_viewport = False
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Adjust_Colour_And_Material'].show_viewport = True
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Write F_DC_And_Merge'].show_viewport = False
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Adjust_Colour_And_Material'].show_render = True
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Write F_DC_And_Merge'].show_render = False
-        bpy.context.scene.sna_dgs_scene_properties.update_mode = 'Disable Camera Updates'
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_50'] = 1
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN'].show_on_cage = (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces')
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN'].show_in_editmode = (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces')
-        bpy.context.view_layer.objects.active.sna_dgs_object_properties.cam_update = False
-        bpy.context.view_layer.objects.active.sna_dgs_object_properties.update_mode = 'Disable Camera Updates'
-        if property_exists("bpy.data.materials['KIRI_3DGS_Render_Material']", globals(), locals()):
-            pass
-        else:
-            before_data = list(bpy.data.materials)
-            bpy.ops.wm.append(directory=os.path.join(os.path.dirname(__file__), 'assets', '3DGS Render APPEND V4.blend') + r'\Material', filename='KIRI_3DGS_Render_Material', link=False)
-            new_data = list(filter(lambda d: not d in before_data, list(bpy.data.materials)))
-            appended_C1470 = None if not new_data else new_data[0]
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Sorter_GN'].show_viewport = (bpy.data.materials['KIRI_3DGS_Render_Material'].surface_render_method == 'BLENDED')
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Sorter_GN'].show_render = (bpy.data.materials['KIRI_3DGS_Render_Material'].surface_render_method == 'BLENDED')
-        input_object = bpy.context.view_layer.objects.active
-        material_name = 'KIRI_3DGS_Render_Material'
-        # Input Variable Names
-        #input_object = None  # Should be set to a bpy.types.Object pointer before running
-        #material_name = "KIRI_3DGS_Render_Material"  # Name of the material to assign
-        # Check if the input object is provided and is valid
-        if not input_object or input_object.type != 'MESH':
-            print("Error: No valid mesh object provided as input.")
-        else:
-            # Get the object and its mesh data
-            obj = input_object
-            mesh = obj.data
-            try:
-                # Remove all existing material slots
-                while len(obj.material_slots) > 0:
-                    bpy.context.object.active_material_index = 0  # Set to the first slot to remove
-                    bpy.ops.object.material_slot_remove()
-                # Check if the material exists; create it if it doesn’t
-                if material_name not in bpy.data.materials:
-                    new_material = bpy.data.materials.new(name=material_name)
-                    new_material.use_nodes = True  # Enable node-based shading (optional, matching your original script)
-                else:
-                    new_material = bpy.data.materials[material_name]
-                # Add the material to the object as a new slot
-                obj.data.materials.append(new_material)
-                print(f"Assigned material '{material_name}' to {obj.name} and removed existing material slots.")
-            except Exception as e:
-                print(f"Error assigning material to {obj.name}: {e}")
-        bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_61'] = bpy.data.materials['KIRI_3DGS_Render_Material']
-        bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
-        if bpy.context and bpy.context.screen:
-            for a in bpy.context.screen.areas:
-                a.tag_redraw()
-        if (bpy.context.scene.sna_dgs_scene_properties.import_uv and (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces')):
-            bpy.ops.object.mode_set('INVOKE_DEFAULT', mode='EDIT')
-            bpy.ops.mesh.select_all('INVOKE_DEFAULT', action='SELECT')
-            bpy.ops.mesh.dissolve_limited('INVOKE_DEFAULT', )
-            bpy.ops.uv.reset('INVOKE_DEFAULT', )
-            bpy.ops.object.mode_set('INVOKE_DEFAULT', mode='OBJECT')
-        if bpy.context.scene.sna_dgs_scene_properties.import_proxy:
-            sna_b2_load_from_blender_object_F0CCB(bpy.context.view_layer.objects.active.name + 'Splat_Proxy')
-        return {"FINISHED"}
+# class SNA_OT_Dgs_Render_Import_Ply_E0A3A(bpy.types.Operator, ImportHelper):
+#     bl_idname = "sna.dgs_render_import_ply_e0a3a"
+#     bl_label = "3DGS Render: Import PLY"
+#     bl_description = "Imports a .PLY file and adds 3DGS modifiers and attributes."
+#     bl_options = {"REGISTER", "UNDO"}
+#     filter_glob: bpy.props.StringProperty( default='*.ply', options={'HIDDEN'} )
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         ply_import_path = self.filepath
+#         output_object = None
+#         attributes_missing = None
+#         message = None
+#         import os
+#         # Input file path
+#         #ply_import_path = "C:\\Users\\joe and pig\\Documents\\Flamingo.ply"  # Example path; update as needed
+#         # Check if the file path is provided and exists
+#         if not ply_import_path or not os.path.exists(ply_import_path):
+#             print("Error: No file path provided or file not found.")
+#             attributes_missing = True
+#             message = f"Error: No file path provided or file not found at {ply_import_path}."
+#             output_object = None
+#         else:
+#             try:
+#                 # Import the PLY file using Blender's native importer (Blender 4.0+)
+#                 try:
+#                     bpy.ops.wm.ply_import(filepath=ply_import_path)
+#                 except AttributeError:
+#                     print("Error: PLY importer operator 'bpy.ops.wm.ply_import' not found. Ensure Blender 4.0 or later is used.")
+#                     attributes_missing = True
+#                     message = f"Error: PLY importer operator not found for {os.path.basename(ply_import_path)}. Ensure Blender 4.0 or later is used."
+#                     output_object = None
+#                 else:
+#                     # Get the last imported object (assuming it's the PLY mesh)
+#                     imported_objects = [obj for obj in bpy.context.scene.objects if obj.select_get()]
+#                     if not imported_objects:
+#                         print("Error: No objects were imported from the PLY file.")
+#                         attributes_missing = True
+#                         message = f"Error: No objects were imported from the PLY file {os.path.basename(ply_import_path)}."
+#                         output_object = None
+#                     else:
+#                         # Use the first imported object (assuming it's a mesh)
+#                         obj = imported_objects[0]
+#                         if obj.type != 'MESH':
+#                             print("Error: Imported object is not a mesh.")
+#                             attributes_missing = True
+#                             message = f"Error: Imported object from {os.path.basename(ply_import_path)} is not a mesh."
+#                             output_object = None
+#                         else:
+#                             # Get the mesh data
+#                             mesh = obj.data
+#                             # List of required 3DGS attributes
+#                             required_attributes = ['f_dc_0', 'f_dc_1', 'f_dc_2', 'opacity', 'scale_0', 'scale_1', 'scale_2', 'rot_0', 'rot_1', 'rot_2', 'rot_3', 'f_rest_0']
+#                             # Check if all required attributes exist on the mesh
+#                             attributes_missing = False
+#                             missing_attrs = []
+#                             for attr_name in required_attributes:
+#                                 if attr_name not in mesh.attributes:
+#                                     attributes_missing = True
+#                                     missing_attrs.append(attr_name)
+#                             # Set the message and output object based on success or failure
+#                             if attributes_missing:
+#                                 message = f"Error: One or more required 3DGS attributes ({', '.join(missing_attrs)}) are missing on {obj.name}. Check that .ply is a 3DGS scan."
+#                                 print(message)
+#                                 output_object = None
+#                             else:
+#                                 message = f"Successfully imported PLY file {os.path.basename(ply_import_path)} and all required 3DGS attributes ({', '.join(required_attributes)}) are present on {obj.name}."
+#                                 print(message)
+#                                 output_object = obj
+#             except Exception as e:
+#                 print(f"Error importing PLY file: {e}")
+#                 attributes_missing = True
+#                 message = f"Error importing PLY file {os.path.basename(ply_import_path)}: {str(e)}"
+#                 output_object = None
+#         # The variables are now available for use
+#         # You can access 'attributes_missing' (boolean), 'message' (string), and 'output_object' (bpy.types.Object or None) in other scripts or logic
+#         if (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces'):
+#             bpy.context.view_layer.objects.active['3DGS_Mesh_Type'] = 'face'
+#             sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Store_Origpos_GN', 'KIRI_3DGS_Store_Origpos_GN', bpy.context.view_layer.objects.active)
+#             bpy.ops.object.modifier_apply('INVOKE_DEFAULT', modifier='KIRI_3DGS_Store_Origpos_GN')
+#             sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Render_GN', 'KIRI_3DGS_Render_GN', bpy.context.view_layer.objects.active)
+#             sna_align_active_values_to_x_4CE1F()
+#             bpy.ops.object.modifier_apply('INVOKE_DEFAULT', modifier='KIRI_3DGS_Render_GN')
+#         else:
+#             bpy.context.view_layer.objects.active['3DGS_Mesh_Type'] = 'vert'
+#         import numpy as np
+#         # Constants from your script
+#         SH_0 = 0.28209479177387814
+#         # Get the active object
+#         obj = bpy.context.active_object
+#         if not obj or obj.type != 'MESH':
+#             print("Error: No active mesh object selected.")
+#         else:
+#             # Get the mesh data
+#             mesh = obj.data
+#             # Check if the required attributes exist
+#             if not all(attr.name in mesh.attributes for attr in mesh.attributes if attr.name in ['f_dc_0', 'f_dc_1', 'f_dc_2', 'opacity']):
+#                 print("Error: Required attributes (f_dc_0, f_dc_1, f_dc_2, opacity) not found on the mesh.")
+#             else:
+#                 # Get the number of points (vertices)
+#                 point_count = len(mesh.vertices)
+#                 expected_length = point_count * 4  # Each point has 1 RGBA set (4 values) for both Col and KIRI_3DGS_Paint
+#                 # Extract data from attributes (assuming they are on POINT domain)
+#                 f_dc_0_data = np.array([v.value for v in mesh.attributes['f_dc_0'].data])
+#                 f_dc_1_data = np.array([v.value for v in mesh.attributes['f_dc_1'].data])
+#                 f_dc_2_data = np.array([v.value for v in mesh.attributes['f_dc_2'].data])
+#                 opacity_data = np.array([v.value for v in mesh.attributes['opacity'].data])
+#                 # Debug: Check lengths of input attribute data
+#                 print(f"Number of points (vertices): {point_count}")
+#                 print(f"Length of f_dc_0_data: {len(f_dc_0_data)}")
+#                 print(f"Length of f_dc_1_data: {len(f_dc_1_data)}")
+#                 print(f"Length of f_dc_2_data: {len(f_dc_2_data)}")
+#                 print(f"Length of opacity_data: {len(opacity_data)}")
+#                 # Verify that attribute data lengths match the number of points
+#                 if not (len(f_dc_0_data) == len(f_dc_1_data) == len(f_dc_2_data) == len(opacity_data) == point_count):
+#                     print("Error: Mismatch in attribute data lengths. Expected length matches point_count.")
+#                 else:
+#                     # Calculate RGB and Alpha for each point (same calculation for both Col and KIRI_3DGS_Paint)
+#                     color_data = []  # For both Col and KIRI_3DGS_Paint, one RGBA per point
+#                     for i in range(point_count):
+#                         # Calculate RGB (matching your script)
+#                         R = (f_dc_0_data[i] * SH_0 + 0.5)
+#                         G = (f_dc_1_data[i] * SH_0 + 0.5)
+#                         B = (f_dc_2_data[i] * SH_0 + 0.5)
+#                         # Calculate Alpha (using sigmoid if opacity is in log-space, or raw if [0, 1])
+#                         # Here, we assume opacity is in log-space (logits) as in your script
+#                         log_opacity = opacity_data[i]
+#                         A = 1 / (1 + np.exp(-log_opacity))
+#                         # Ensure values are in [0, 1]
+#                         R = max(0.0, min(1.0, R))
+#                         G = max(0.0, min(1.0, G))
+#                         B = max(0.0, min(1.0, B))
+#                         A = max(0.0, min(1.0, A))
+#                         # Add RGBA for both Col and KIRI_3DGS_Paint (one set per point)
+#                         color_data.extend([R, G, B, A])
+#                     # Debug: Check calculated data length
+#                     print(f"Length of color_data (Col and KIRI_3DGS_Paint): {len(color_data)}")
+#                     print(f"Expected length for POINT domain: {expected_length}")
+#                     # Verify data length matches expectation
+#                     if len(color_data) != expected_length:
+#                         print(f"Error: Array length mismatch (expected {expected_length}, got {len(color_data)})")
+#                     else:
+#                         # Create or update the Col attribute on the point domain
+#                         if 'Col' in mesh.attributes:
+#                             mesh.attributes.remove(mesh.attributes['Col'])
+#                         col_attr = mesh.attributes.new(name="Col", type='FLOAT_COLOR', domain='POINT')
+#                         col_attr.data.foreach_set("color", color_data)
+#                         # Create or update the KIRI_3DGS_Paint attribute on the point domain
+#                         if 'KIRI_3DGS_Paint' in mesh.attributes:
+#                             mesh.attributes.remove(mesh.attributes['KIRI_3DGS_Paint'])
+#                         paint_attr = mesh.attributes.new(name="KIRI_3DGS_Paint", type='FLOAT_COLOR', domain='POINT')
+#                         paint_attr.data.foreach_set("color", color_data)
+#                         # Set KIRI_3DGS_Paint as the active color attribute
+#                         mesh.color_attributes.active_color = paint_attr
+#                         print(f"Created Col attribute on {obj.name} with {point_count} points.")
+#                         print(f"Created KIRI_3DGS_Paint attribute on {obj.name} with {point_count} color values on the POINT domain.")
+#         sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Render_GN', 'KIRI_3DGS_Render_GN', bpy.context.view_layer.objects.active)
+#         sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Sorter_GN', 'KIRI_3DGS_Sorter_GN', bpy.context.view_layer.objects.active)
+#         sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Adjust_Colour_And_Material', 'KIRI_3DGS_Adjust_Colour_And_Material', bpy.context.view_layer.objects.active)
+#         sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Write F_DC_And_Merge', 'KIRI_3DGS_Write F_DC_And_Merge', bpy.context.view_layer.objects.active)
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN'].show_viewport = False
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Adjust_Colour_And_Material'].show_viewport = True
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Write F_DC_And_Merge'].show_viewport = False
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Adjust_Colour_And_Material'].show_render = True
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Write F_DC_And_Merge'].show_render = False
+#         bpy.context.scene.sna_dgs_scene_properties.update_mode = 'Disable Camera Updates'
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_50'] = 1
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN'].show_on_cage = (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces')
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN'].show_in_editmode = (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces')
+#         bpy.context.view_layer.objects.active.sna_dgs_object_properties.cam_update = False
+#         bpy.context.view_layer.objects.active.sna_dgs_object_properties.update_mode = 'Disable Camera Updates'
+#         if property_exists("bpy.data.materials['KIRI_3DGS_Render_Material']", globals(), locals()):
+#             pass
+#         else:
+#             before_data = list(bpy.data.materials)
+#             bpy.ops.wm.append(directory=os.path.join(os.path.dirname(__file__), 'assets', '3DGS Render APPEND V4.blend') + r'\Material', filename='KIRI_3DGS_Render_Material', link=False)
+#             new_data = list(filter(lambda d: not d in before_data, list(bpy.data.materials)))
+#             appended_C1470 = None if not new_data else new_data[0]
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Sorter_GN'].show_viewport = (bpy.data.materials['KIRI_3DGS_Render_Material'].surface_render_method == 'BLENDED')
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Sorter_GN'].show_render = (bpy.data.materials['KIRI_3DGS_Render_Material'].surface_render_method == 'BLENDED')
+#         input_object = bpy.context.view_layer.objects.active
+#         material_name = 'KIRI_3DGS_Render_Material'
+#         # Input Variable Names
+#         #input_object = None  # Should be set to a bpy.types.Object pointer before running
+#         #material_name = "KIRI_3DGS_Render_Material"  # Name of the material to assign
+#         # Check if the input object is provided and is valid
+#         if not input_object or input_object.type != 'MESH':
+#             print("Error: No valid mesh object provided as input.")
+#         else:
+#             # Get the object and its mesh data
+#             obj = input_object
+#             mesh = obj.data
+#             try:
+#                 # Remove all existing material slots
+#                 while len(obj.material_slots) > 0:
+#                     bpy.context.object.active_material_index = 0  # Set to the first slot to remove
+#                     bpy.ops.object.material_slot_remove()
+#                 # Check if the material exists; create it if it doesn’t
+#                 if material_name not in bpy.data.materials:
+#                     new_material = bpy.data.materials.new(name=material_name)
+#                     new_material.use_nodes = True  # Enable node-based shading (optional, matching your original script)
+#                 else:
+#                     new_material = bpy.data.materials[material_name]
+#                 # Add the material to the object as a new slot
+#                 obj.data.materials.append(new_material)
+#                 print(f"Assigned material '{material_name}' to {obj.name} and removed existing material slots.")
+#             except Exception as e:
+#                 print(f"Error assigning material to {obj.name}: {e}")
+#         bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_61'] = bpy.data.materials['KIRI_3DGS_Render_Material']
+#         bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
+#         if bpy.context and bpy.context.screen:
+#             for a in bpy.context.screen.areas:
+#                 a.tag_redraw()
+#         if (bpy.context.scene.sna_dgs_scene_properties.import_uv and (bpy.context.scene.sna_dgs_scene_properties.import_face_vert == 'Faces')):
+#             bpy.ops.object.mode_set('INVOKE_DEFAULT', mode='EDIT')
+#             bpy.ops.mesh.select_all('INVOKE_DEFAULT', action='SELECT')
+#             bpy.ops.mesh.dissolve_limited('INVOKE_DEFAULT', )
+#             bpy.ops.uv.reset('INVOKE_DEFAULT', )
+#             bpy.ops.object.mode_set('INVOKE_DEFAULT', mode='OBJECT')
+#         if bpy.context.scene.sna_dgs_scene_properties.import_proxy:
+#             sna_b2_load_from_blender_object_F0CCB(bpy.context.view_layer.objects.active.name + 'Splat_Proxy')
+#         return {"FINISHED"}
 
 
 # def sna_append_and_add_geo_nodes_function_execute_6BCD7(Node_Group_Name, Modifier_Name, Object):
@@ -5068,30 +5068,30 @@ class SNA_OT_Dgs_Render_Import_Ply_E0A3A(bpy.types.Operator, ImportHelper):
 #     return modifier_6D624
 
 
-class SNA_PT_DGS_RENDER_BY_KIRI_ENGINE_A02CB(bpy.types.Panel):
-    bl_label = '3DGS Render by KIRI Engine'
-    bl_idname = 'SNA_PT_DGS_RENDER_BY_KIRI_ENGINE_A02CB'
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_context = ''
-    bl_category = '3DGS Render'
-    bl_order = 0
-    bl_ui_units_x=0
-
-    @classmethod
-    def poll(cls, context):
-        return not (False)
-
-    def draw_header(self, context):
-        layout = self.layout
-        layout.template_icon(icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'kiriengine icon.svg')), scale=1.0)
-
-    def draw(self, context):
-        layout = self.layout
-        layout_function = layout
-        sna_dgs_render__main_function_menu_019C7(layout_function, )
-        layout_function = layout
-        sna_about_kiri_links_docs_3dgs_D02EC(layout_function, )
+# class SNA_PT_DGS_RENDER_BY_KIRI_ENGINE_A02CB(bpy.types.Panel):
+#     bl_label = '3DGS Render by KIRI Engine'
+#     bl_idname = 'SNA_PT_DGS_RENDER_BY_KIRI_ENGINE_A02CB'
+#     bl_space_type = 'VIEW_3D'
+#     bl_region_type = 'UI'
+#     bl_context = ''
+#     bl_category = '3DGS Render'
+#     bl_order = 0
+#     bl_ui_units_x=0
+#
+#     @classmethod
+#     def poll(cls, context):
+#         return not (False)
+#
+#     def draw_header(self, context):
+#         layout = self.layout
+#         layout.template_icon(icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'kiriengine icon.svg')), scale=1.0)
+#
+#     def draw(self, context):
+#         layout = self.layout
+#         layout_function = layout
+#         sna_dgs_render__main_function_menu_019C7(layout_function, )
+#         layout_function = layout
+#         sna_about_kiri_links_docs_3dgs_D02EC(layout_function, )
 
 
 # def sna_modify_menu_AEA26(layout_function, ):
@@ -6066,111 +6066,111 @@ class SNA_PT_DGS_RENDER_BY_KIRI_ENGINE_A02CB(bpy.types.Panel):
 #         op.sna_create_duplicate_and_remove_other_modifiers = False
 #
 
-class SNA_OT_Dgs_Render_Append_Rough_Mesh_Modifier_65Da3(bpy.types.Operator):
-    bl_idname = "sna.dgs_render_append_rough_mesh_modifier_65da3"
-    bl_label = "3DGS Render: Append Rough Mesh Modifier"
-    bl_description = "Adds a Rough Mesh modifier to the active object."
-    bl_options = {"REGISTER", "UNDO"}
-    sna_create_duplicate_and_remove_other_modifiers: bpy.props.BoolProperty(name='Create duplicate and remove other modifiers', description='', options={'HIDDEN'}, default=True)
-
-    @classmethod
-    def poll(cls, context):
-        if bpy.app.version >= (3, 0, 0) and True:
-            cls.poll_message_set('')
-        return not False
-
-    def execute(self, context):
-        if (bpy.context.view_layer.objects.active.type == 'MESH' or bpy.context.view_layer.objects.active.type == 'CURVE'):
-            if self.sna_create_duplicate_and_remove_other_modifiers:
-                source_obj_name = bpy.context.view_layer.objects.active.name
-                # Input variables
-                #source_obj_name = "Cube"  # Change this to your object's name
-                offset_x = 0.0  # Input float variable for X offset
-                # Get the source object
-                source_obj = bpy.data.objects.get(source_obj_name)
-                # Check if the object exists
-                if source_obj:
-                    # Create a copy of the object
-                    new_obj = source_obj.copy()
-                    new_obj.data = source_obj.data.copy()
-                    # Link the new object to the scene
-                    bpy.context.scene.collection.objects.link(new_obj)
-                    # Apply the offset if any
-                    new_obj.location.x += offset_x
-                    # Clear current selection
-                    bpy.ops.object.select_all(action='DESELECT')
-                    # Select and activate the new object
-                    new_obj.select_set(True)
-                    bpy.context.view_layer.objects.active = new_obj
-                    # Store the new object's name in a variable
-                    new_object_name = new_obj.name
-                    # Output the new object for Serpens (return the actual object)
-                    output_object = new_obj
-                else:
-                    new_object_name = "ERROR: Source object not found"
-                    output_object = None
-                # Output the new object's name (this will be captured by Serpens)
-                print(new_object_name)
-                if (property_exists("bpy.context.view_layer.objects.active.modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.view_layer.objects.active.modifiers):
-                    bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_50'] = 1
-                    bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
-                    if bpy.context and bpy.context.screen:
-                        for a in bpy.context.screen.areas:
-                            a.tag_redraw()
-                    object_name = bpy.context.view_layer.objects.active.name
-                    # Replace this with your object's name
-                    #object_name = "YourObjectName"
-                    # Get the object
-                    obj = bpy.data.objects.get(object_name)
-                    if obj:
-                        # Make sure the object is selected and active
-                        bpy.context.view_layer.objects.active = obj
-                        obj.select_set(True)
-                        # Apply all modifiers
-                        for modifier in obj.modifiers[:]:  # [:] creates a copy of the list to avoid modification issues
-                            try:
-                                bpy.ops.object.modifier_apply(modifier=modifier.name)
-                                print(f"Applied modifier: {modifier.name}")
-                            except Exception as e:
-                                print(f"Failed to apply modifier {modifier.name}: {str(e)}")
-                    else:
-                        print(f"Object '{object_name}' not found")
-            else:
-                if (property_exists("bpy.context.view_layer.objects.active.modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.view_layer.objects.active.modifiers):
-                    bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_50'] = 1
-                    bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
-                    if bpy.context and bpy.context.screen:
-                        for a in bpy.context.screen.areas:
-                            a.tag_redraw()
-            created_modifier_0_827c5 = sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Convert_To_Rough_Mesh_GN', 'KIRI_3DGS_Convert_To_Rough_Mesh_GN', bpy.context.view_layer.objects.active)
-            for i_1EB8C in range(len(bpy.context.view_layer.objects.active.modifiers)):
-                if (bpy.context.view_layer.objects.active.modifiers[i_1EB8C] == created_modifier_0_827c5):
-                    bpy.context.view_layer.objects.active.modifiers.move(from_index=i_1EB8C, to_index=0, )
-                    bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
-                    if bpy.context and bpy.context.screen:
-                        for a in bpy.context.screen.areas:
-                            a.tag_redraw()
-        else:
-            self.report({'INFO'}, message='The Active Object is not a mesh or curve object.')
-        return {"FINISHED"}
-
-    def draw(self, context):
-        layout = self.layout
-        box_F6B46 = layout.box()
-        box_F6B46.alert = False
-        box_F6B46.enabled = True
-        box_F6B46.active = True
-        box_F6B46.use_property_split = False
-        box_F6B46.use_property_decorate = False
-        box_F6B46.alignment = 'Expand'.upper()
-        box_F6B46.scale_x = 1.0
-        box_F6B46.scale_y = 1.0
-        if not True: box_F6B46.operator_context = "EXEC_DEFAULT"
-        box_F6B46.prop(self, 'sna_create_duplicate_and_remove_other_modifiers', text='Create duplicate and remove other modifiers', icon_value=0, emboss=True)
-        box_F6B46.label(text='The 3DGS Render modifier will have Camera Updates disabled', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'tips-one.svg')))
-
-    def invoke(self, context, event):
-        return context.window_manager.invoke_props_dialog(self, width=400)
+# class SNA_OT_Dgs_Render_Append_Rough_Mesh_Modifier_65Da3(bpy.types.Operator):
+#     bl_idname = "sna.dgs_render_append_rough_mesh_modifier_65da3"
+#     bl_label = "3DGS Render: Append Rough Mesh Modifier"
+#     bl_description = "Adds a Rough Mesh modifier to the active object."
+#     bl_options = {"REGISTER", "UNDO"}
+#     sna_create_duplicate_and_remove_other_modifiers: bpy.props.BoolProperty(name='Create duplicate and remove other modifiers', description='', options={'HIDDEN'}, default=True)
+#
+#     @classmethod
+#     def poll(cls, context):
+#         if bpy.app.version >= (3, 0, 0) and True:
+#             cls.poll_message_set('')
+#         return not False
+#
+#     def execute(self, context):
+#         if (bpy.context.view_layer.objects.active.type == 'MESH' or bpy.context.view_layer.objects.active.type == 'CURVE'):
+#             if self.sna_create_duplicate_and_remove_other_modifiers:
+#                 source_obj_name = bpy.context.view_layer.objects.active.name
+#                 # Input variables
+#                 #source_obj_name = "Cube"  # Change this to your object's name
+#                 offset_x = 0.0  # Input float variable for X offset
+#                 # Get the source object
+#                 source_obj = bpy.data.objects.get(source_obj_name)
+#                 # Check if the object exists
+#                 if source_obj:
+#                     # Create a copy of the object
+#                     new_obj = source_obj.copy()
+#                     new_obj.data = source_obj.data.copy()
+#                     # Link the new object to the scene
+#                     bpy.context.scene.collection.objects.link(new_obj)
+#                     # Apply the offset if any
+#                     new_obj.location.x += offset_x
+#                     # Clear current selection
+#                     bpy.ops.object.select_all(action='DESELECT')
+#                     # Select and activate the new object
+#                     new_obj.select_set(True)
+#                     bpy.context.view_layer.objects.active = new_obj
+#                     # Store the new object's name in a variable
+#                     new_object_name = new_obj.name
+#                     # Output the new object for Serpens (return the actual object)
+#                     output_object = new_obj
+#                 else:
+#                     new_object_name = "ERROR: Source object not found"
+#                     output_object = None
+#                 # Output the new object's name (this will be captured by Serpens)
+#                 print(new_object_name)
+#                 if (property_exists("bpy.context.view_layer.objects.active.modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.view_layer.objects.active.modifiers):
+#                     bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_50'] = 1
+#                     bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
+#                     if bpy.context and bpy.context.screen:
+#                         for a in bpy.context.screen.areas:
+#                             a.tag_redraw()
+#                     object_name = bpy.context.view_layer.objects.active.name
+#                     # Replace this with your object's name
+#                     #object_name = "YourObjectName"
+#                     # Get the object
+#                     obj = bpy.data.objects.get(object_name)
+#                     if obj:
+#                         # Make sure the object is selected and active
+#                         bpy.context.view_layer.objects.active = obj
+#                         obj.select_set(True)
+#                         # Apply all modifiers
+#                         for modifier in obj.modifiers[:]:  # [:] creates a copy of the list to avoid modification issues
+#                             try:
+#                                 bpy.ops.object.modifier_apply(modifier=modifier.name)
+#                                 print(f"Applied modifier: {modifier.name}")
+#                             except Exception as e:
+#                                 print(f"Failed to apply modifier {modifier.name}: {str(e)}")
+#                     else:
+#                         print(f"Object '{object_name}' not found")
+#             else:
+#                 if (property_exists("bpy.context.view_layer.objects.active.modifiers", globals(), locals()) and 'KIRI_3DGS_Render_GN' in bpy.context.view_layer.objects.active.modifiers):
+#                     bpy.context.view_layer.objects.active.modifiers['KIRI_3DGS_Render_GN']['Socket_50'] = 1
+#                     bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
+#                     if bpy.context and bpy.context.screen:
+#                         for a in bpy.context.screen.areas:
+#                             a.tag_redraw()
+#             created_modifier_0_827c5 = sna_append_and_add_geo_nodes_function_execute_6BCD7('KIRI_3DGS_Convert_To_Rough_Mesh_GN', 'KIRI_3DGS_Convert_To_Rough_Mesh_GN', bpy.context.view_layer.objects.active)
+#             for i_1EB8C in range(len(bpy.context.view_layer.objects.active.modifiers)):
+#                 if (bpy.context.view_layer.objects.active.modifiers[i_1EB8C] == created_modifier_0_827c5):
+#                     bpy.context.view_layer.objects.active.modifiers.move(from_index=i_1EB8C, to_index=0, )
+#                     bpy.context.view_layer.objects.active.update_tag(refresh={'DATA'}, )
+#                     if bpy.context and bpy.context.screen:
+#                         for a in bpy.context.screen.areas:
+#                             a.tag_redraw()
+#         else:
+#             self.report({'INFO'}, message='The Active Object is not a mesh or curve object.')
+#         return {"FINISHED"}
+#
+#     def draw(self, context):
+#         layout = self.layout
+#         box_F6B46 = layout.box()
+#         box_F6B46.alert = False
+#         box_F6B46.enabled = True
+#         box_F6B46.active = True
+#         box_F6B46.use_property_split = False
+#         box_F6B46.use_property_decorate = False
+#         box_F6B46.alignment = 'Expand'.upper()
+#         box_F6B46.scale_x = 1.0
+#         box_F6B46.scale_y = 1.0
+#         if not True: box_F6B46.operator_context = "EXEC_DEFAULT"
+#         box_F6B46.prop(self, 'sna_create_duplicate_and_remove_other_modifiers', text='Create duplicate and remove other modifiers', icon_value=0, emboss=True)
+#         box_F6B46.label(text='The 3DGS Render modifier will have Camera Updates disabled', icon_value=load_preview_icon(os.path.join(os.path.dirname(__file__), 'assets', 'tips-one.svg')))
+#
+#     def invoke(self, context, event):
+#         return context.window_manager.invoke_props_dialog(self, width=400)
 
 # def sna_crop_box_F2C60(layout_function, ):
 #     if (property_exists("bpy.context.view_layer.objects.active.modifiers", globals(), locals()) and 'KIRI_3DGS_Crop_Box_GN' in bpy.context.view_layer.objects.active.modifiers):
